Source maps.
Enables source maps to see the real source code while debugging in the browser console.
We use `source-map-inline` instead of `source-map` because the browser caches the `.map` files, so it doesn't update with hot reload.
Rendering the map inline actually appends the source map directly on the compiled file, so hot reload works just fine.