Loading images.
Uses the `url-loader` to load image files. This loader in particular has the `limit` option which automatically converts
any parsed file -with size less than the specified (in bytes)- into a base64 data string.

We could also have used the simpler `file-loader` or the more specialized `image-webpack-loader`, it dependends on
the app requirements.